--- kinderboekenapp/.env
@@ -1,1 +1,0 @@
-CHOKIDAR_USEPOLLING=true
