--- kinderboekenapp/src/Mascotte.css
@@ -1,11 +1,0 @@
-.mascotte-name__section {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-.mascotte-name__section__figure__img {
-  height: 25rem;
-}
-
-/*# sourceMappingURL=Mascotte.css.map */
