--- kinderboekenapp/src/Modal.css
@@ -1,43 +1,0 @@
-.modal {
-  position: fixed;
-  background: white;
-  width: 80%;
-  height: 70vh;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-.modal__header {
-  height: 10%;
-  background-color: #00aeff;
-  padding: 0 0.75rem;
-}
-.modal__header__close {
-  float: right;
-  font-size: 4rem;
-  color: white;
-}
-.modal__header__h2 {
-  font-size: 1.6rem;
-  padding-top: 1.75rem;
-  color: white;
-}
-
-.popup {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.6);
-}
-
-.display-block {
-  display: block;
-}
-
-.display-none {
-  display: none;
-}
-
-/*# sourceMappingURL=Modal.css.map */
