--- kinderboekenapp/src/Searchbar.css
@@ -1,15 +1,0 @@
-.searchbar_section{
-    width: 80%; 
-    margin: 8%;
-    margin-bottom: 2rem;    
-}
-
-.searchbar_section__input{
-    border-radius: 0.5rem;
-    border: 1px grey solid;
-    width: 100%;
-    padding: 1rem;
-    font-size: 1.4rem;
-    font-family: 'Open Sans', sans-serif;
-   
-}
