+++ kinderboekenapp/src/opdrachten/BookAssignments.js
@@ -5,10 +5,8 @@ import axios from "axios";
 import OpdrachtNav from './components/OpdrachtNav';
 import authHeader from '../services/service_auth-header';
 import { changeUserAssignments, changeActiveBook } from "../actions";
-import ImageVraag from './ImageVraag';
-import Endscreen from './components/Endscreen.js';
 
-import { AudioVraag, ColorVraag, ColorAntwoord, ImageAntwoord,  MultipleChoice, Podcast, VraagUnavailable, MultipleChoice2} from "./Vragen"
+import { ImageVraag, AudioVraag, ColorVraag, ColorAntwoord, ImageAntwoord,  MultipleChoice, Podcast, VraagUnavailable, MultipleChoice2} from "./Vragen"
 import "./BookAssignments.css";
 
 class BookAssignments extends Component {
@@ -101,7 +99,6 @@ class BookAssignments extends Component {
                     <OpdrachtNav />
                     <a className="back-link" href='/boekenplank'><i className="fas fa-times"></i></a>
                     {this.props.userAssignments.map((assignment, index) => this.createAssignments(assignment, index))}
-                    <Endscreen/>
                 </article>
             );
         }
