+++ kinderboekenapp/src/opdrachten/ImageVraag.css
@@ -43,10 +43,4 @@
   width: 100%;
 }
 
-@media screen and (max-width: 60rem) {
-  .assignment-container {
-    font-size: 1.5rem;
-  }
-}
-
 /*# sourceMappingURL=ImageVraag.css.map */
