+++ kinderboekenapp/src/opdrachten/ImageVraag.scss
@@ -43,9 +43,3 @@
 .image{
     width: 100%;
 }
-
-@media screen and (max-width: 60rem){
-    .assignment-container{
-      font-size: 1.5rem;
-    }
-  }
