--- kinderboekenapp/src/opdrachten/VraagImage.css
@@ -1,34 +1,0 @@
-.img-container {
-  max-height: 8rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.img-container-imageVraag {
-  transform: scale(0.2);
-}
-
-.custom-file-input::-webkit-file-upload-button {
-  visibility: hidden;
-}
-
-.custom-file-input::before {
-  content: "Kies een foto";
-  display: inline-block;
-  background: linear-gradient(top, #f9f9f9, #e3e3e3);
-  border: 0.1rem solid #999;
-  border-radius: 0.3rem;
-  padding: 0.5rem 0.8rem;
-  outline: none;
-  white-space: nowrap;
-  cursor: pointer;
-  text-shadow: 0.1rem 0.1rem #fff;
-  font-size: 1.6rem;
-}
-@media screen and (min-width: 600px){
-  .assignment-container{
-    font-size: 15px;
-    font-weight: 200;
-  }
-}
-/*# sourceMappingURL=VraagImage.css.map */
