--- kinderboekenapp/src/opdrachten/VraagImage.scss
@@ -1,29 +1,0 @@
-.img-container {
-    // max-width: 40%;
-    max-height: 6rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    &-imageVraag {
-        transform: scale(0.1);
-    }
-}
-
-.custom-file-input::-webkit-file-upload-button {
-    visibility: hidden;
-  }
-
-
-.custom-file-input::before {
-    content: 'Kies een foto';
-    display: inline-block;
-    background: linear-gradient(top, #f9f9f9, #e3e3e3);
-    border: 0.1rem solid #999;
-    border-radius: 0.3rem;
-    padding: 0.5rem 0.8rem;
-    outline: none;
-    white-space: nowrap;
-    cursor: pointer;
-    text-shadow: 0.1rem 0.1rem #fff;
-    font-size: 1.6rem;
-}
