+++ kinderboekenapp/src/profile/ColorOptions.scss
@@ -45,7 +45,13 @@
 
 @media screen and (min-width: 768px){
     .colorForm__btnSection__submitButton {
-        margin-top: 20rem;
+        margin-top: 30rem;
         font-size: 3rem;
+    }
+}   
+
+@media screen and (min-width: 1000px){
+    .colorForm__btnSection__submitButton {
+        margin-top: 50rem;
     }
 }
