+++ kinderboekenapp/src/profile/Modal.css
@@ -62,7 +62,6 @@
   }
   .modal__header__h2 {
     font-size: 2.8rem;
-    padding-top: 2rem;
   }
 }
