--- package.json
@@ -1,22 +1,0 @@
-{
-  "name": "ipmedt4-kinderboekenapp-react",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Baeshee/IPMEDT4-Kinderboekenapp-React.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/Baeshee/IPMEDT4-Kinderboekenapp-React/issues"
-  },
-  "homepage": "https://github.com/Baeshee/IPMEDT4-Kinderboekenapp-React#readme",
-  "dependencies": {
-    "start": "^5.1.0"
-  }
-}
